Add admin to user and only allow admin to edit/delete

+++ app/routes.py
@@ -1,21 +1,14 @@
 from app import app
 from flask import render_template, redirect, url_for, flash
-from flask_login import login_user, logout_user, login_required
+from flask_login import login_user, logout_user, login_required, current_user
 from app.forms import RegisterForm, LoginForm, ProductForm
 from app.models import User, Product, Category
+
 
 @app.route('/')
 def index():
     products = Product.query.all()
     return render_template('index.html', products=products)
-
-
-@app.route('/name')
-@login_required
-def name():
-    my_name = 'Brian'
-
-    return render_template('name.html', name=my_name)
 
 
 @app.route('/register', methods=["GET","POST"])
@@ -84,6 +77,9 @@ def product_info(prod_id):
 @app.route('/products/<int:prod_id>/edit', methods=["GET", "POST"])
 @login_required
 def edit_product(prod_id):
+    if not current_user.is_admin:
+        flash("Excuse me, you are not allowed here.", "warning")
+        return redirect(url_for('index'))
     product = Product.query.get_or_404(prod_id)
     form = ProductForm()
     form.category_id.choices = [(c.id, c.name) for c in Category.query.all()]
@@ -110,7 +106,11 @@ def edit_product(prod_id):
 
 
 @app.route('/products/<int:prod_id>/delete')
+@login_required
 def delete_product(prod_id):
+    if not current_user.is_admin:
+        flash("Excuse me, you are not allowed here.", "warning")
+        return redirect(url_for('index'))
     product = Product.query.get_or_404(prod_id)
     product.delete()
     flash(f'{product.name} has been deleted', 'danger')
